js: use jsbundler-rails with webpack
This removes webpacker, and uses webpack directly.

Not working yet:

- Sourcemaps
- Chunks
- Not recompiling on every watched change
