+++ .eslintrc.js
@@ -26,7 +26,9 @@ module.exports = {
   },
   overrides: [
     {
-      files: ['config/webpack/**/*.js', 'babel.config.js', 'postcss.config.js'],
+      files: [
+        '*.config.js'
+      ],
       env: {
         node: true
       }
