+++ .gitignore
@@ -23,6 +23,8 @@ Procfile.dev
 storage/
 /public/packs
 /public/packs-test
+/app/assets/builds/*
+!/app/assets/builds/.keep
 /node_modules
 /yarn-error.log
 yarn-debug.log*
@@ -34,4 +36,3 @@ yarn-debug.log*
 
 # Local Netlify folder
 .netlify
-
