+++ app/javascript/manager.js
@@ -1,0 +1,1 @@
+import './manager/fields/features';
