--- app/javascript/packs/manager.js
@@ -1,1 +1,0 @@
-import '../manager/fields/features';
