--- app/javascript/packs/track-admin.js
@@ -1,2 +1,0 @@
-import '../shared/track/sendinblue';
-import '../shared/track/crisp';
