--- app/javascript/packs/track.js
@@ -1,2 +1,0 @@
-import '../shared/track/matomo';
-import '../shared/track/sentry';
