+++ app/javascript/track-admin.js
@@ -1,0 +1,2 @@
+import './shared/track/sendinblue';
+import './shared/track/crisp';
