+++ app/javascript/track.js
@@ -1,0 +1,2 @@
+import './shared/track/matomo';
+import './shared/track/sentry';
