+++ app/views/layouts/application.html.haml
@@ -13,8 +13,8 @@
     = favicon_link_tag(image_url("#{FAVICON_32PX_SRC}"), type: "image/png", sizes: "32x32")
     = favicon_link_tag(image_url("#{FAVICON_96PX_SRC}"), type: "image/png", sizes: "96x96")
 
-    - packs = ['application', 'track', administrateur_signed_in? ? 'track-admin' : nil].compact
-    = javascript_packs_with_chunks_tag *packs, defer: true
+    - bundles = ['application', 'track', administrateur_signed_in? ? 'track-admin' : nil].compact
+    = javascript_include_tag *bundles, defer: true
 
     = preload_link_tag(asset_url("Muli-Regular.woff2"))
     = preload_link_tag(asset_url("Muli-Bold.woff2"))
