--- config/webpack/test.js
@@ -1,5 +1,0 @@
-process.env.NODE_ENV = process.env.NODE_ENV || 'development';
-
-const environment = require('./environment');
-
-module.exports = environment.toWebpackConfig();
