+++ package.json
@@ -1,5 +1,8 @@
 {
   "dependencies": {
+    "@babel/core": "^7.16.5",
+    "@babel/plugin-transform-runtime": "^7.16.5",
+    "@babel/preset-env": "^7.16.5",
     "@babel/preset-react": "^7.14.5",
     "@headlessui/react": "^1.3.0",
     "@heroicons/react": "^1.0.1",
@@ -8,16 +11,17 @@
     "@rails/actiontext": "^6.1.4-1",
     "@rails/activestorage": "^6.1.4-1",
     "@rails/ujs": "^6.1.4-1",
-    "@rails/webpacker": "5.4.3",
     "@reach/combobox": "^0.13.0",
     "@reach/slider": "^0.15.0",
     "@reach/visually-hidden": "^0.15.2",
     "@sentry/browser": "6.12.0",
     "@tmcw/togeojson": "^4.3.0",
+    "babel-loader": "^8.2.3",
     "babel-plugin-macros": "^2.8.0",
     "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
     "chartkick": "^3.2.0",
     "core-js": "^3.6.5",
+    "css-loader": "^6.5.1",
     "debounce": "^1.2.1",
     "dom4": "^2.1.6",
     "email-butler": "^1.0.13",
@@ -36,10 +40,11 @@
     "react-popper": "^2.2.5",
     "react-query": "^3.9.7",
     "react-sortable-hoc": "^1.11.0",
+    "style-loader": "^3.3.1",
     "trix": "^1.2.3",
     "use-debounce": "^5.2.0",
-    "webpack": "^4.46.0",
-    "webpack-cli": "^3.3.12",
+    "webpack": "^5.65.0",
+    "webpack-cli": "^4.9.1",
     "whatwg-fetch": "^3.0.0"
   },
   "devDependencies": {
@@ -56,8 +61,9 @@
     "webpack-dev-server": "^4.6.0"
   },
   "scripts": {
-    "lint:js": "eslint --ext .js,.jsx,.ts,.tsx ./app/javascript ./config/webpack",
-    "webpack:build": "NODE_ENV=production bin/webpack",
+    "lint:js": "eslint --ext .js,.jsx,.ts,.tsx ./app/javascript webpack.config.js",
+    "build": "webpack --config webpack.config.js --node-env=production",
+    "build-dev": "webpack --config webpack.config.js",
     "graphql:docs:build": "graphdoc --force",
     "graphql:docs:deploy": "netlify deploy -d ./docs/graphql --prod",
     "graphql:docs:publish": "yarn graphql:docs:build && yarn graphql:docs:deploy"
